feat: ability to specify plugin's version to download
Note that plugin's checksum can't be verified when specifying a version

+++ app/Extensions/UpdateManager.php
@@ -129,7 +129,7 @@ class UpdateManager
         return $updates;
     }
 
-    public function download(array $info, string $tempDir = '')
+    public function download(array $info, string $tempDir = '', string $pluginVersion = 'latest')
     {
         $updatesPath = storage_path('app/updates/');
 
@@ -152,12 +152,19 @@ class UpdateManager
             $this->files->delete($path);
         }
 
+        $downloadUrl = $info['url'];
+        if ($pluginVersion !== 'latest') {
+            $downloadUrl = preg_replace("#\/updates\/\d+\/download$#", "/download/$pluginVersion", $downloadUrl);
+        }
+
         $this->prepareHttpRequest()
             ->withOptions(['sink' => $path])
-            ->get($info['url'])
+            ->get($downloadUrl)
             ->throw();
 
-        if (! hash_equals($info['hash'], hash_file('sha256', $path))) {
+        // TODO : Verify checksum for non-latest plugin's version
+        // Not possible for now because auzriom.com API does not return plugin's hash for older versions.
+        if (! hash_equals($info['hash'], hash_file('sha256', $path)) && $pluginVersion === 'latest') {
             $this->files->delete($path);
 
             throw new RuntimeException('The file hash do not match expected hash!');
